Clases/Clase34/cac_movies_bkn_24/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
mysql==0.0.3
mysqlclient==2.2.0
Pillow==10.1.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
